Change headings
Change 'searching for Git commits applied to a single file' to
'searching for commits applied to a single file' heading.
Change 'searching for Git commits in which a string was added or
removed' heading to 'searching for commits in which patch text was added
or removed' heading.
Change contents of admonitions.
Change unordered list items.