Change headings
Change 'searching a Git repository' heading to 'exploring Git history'
heading.
Change 'searching for commits applied to a single file' heading to
'listing commits applied to a relevant file' heading.
Change 'searching for commits in which patch text was added or removed'
heading to 'listing commits in which relevant text was added to a patch
or removed from a patch' heading.
Change 'searching for text in log messages' heading to 'listing commits
in which relevant text was present in a log message' heading.
Change contents of admonitions.
Change unordered list items.