Add information for 15 table rows, add footnotes, change footnote
Add predominant programming language for 15 table rows.
Add footnote evidencing license for QtQR.
Add footnote evidencing license for SoX.
Change 'VLC media player' footnote to use a URL that is likely to be
more future-proof.
Change 'KRename' hyperlink to a URL that does not produce a detection of
a potential security threat in Firefox.